pyproject.toml
[project]
name = "sentry2cc"
version = "0.1.0"
description = "Sentry to Claude Code – automatically fix Sentry errors with Claude Code Agent"
readme = "README.md"
authors = [
    { name = "chandralegend", email = "irugalbandarachandra@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.0.16",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
sentry2cc = "sentry2cc:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"
